Revert "Export views (#12824)"
This reverts commit 8c2ef27eded51a36fd5c3c9abe4f89221d04e9c0.
