+++ ydb/core/tx/schemeshard/schemeshard_export.cpp
@@ -170,8 +170,7 @@ void TSchemeShard::PersistCreateExport(NIceDb::TNiceDb& db, const TExportInfo::T
             NIceDb::TUpdate<Schema::ExportItems::SourcePathName>(item.SourcePathName),
             NIceDb::TUpdate<Schema::ExportItems::SourceOwnerPathId>(item.SourcePathId.OwnerId),
             NIceDb::TUpdate<Schema::ExportItems::SourcePathId>(item.SourcePathId.LocalPathId),
-            NIceDb::TUpdate<Schema::ExportItems::State>(static_cast<ui8>(item.State)),
-            NIceDb::TUpdate<Schema::ExportItems::SourcePathType>(item.SourcePathType)
+            NIceDb::TUpdate<Schema::ExportItems::State>(static_cast<ui8>(item.State))
         );
     }
 }
@@ -230,10 +229,6 @@ void TSchemeShard::Handle(TEvExport::TEvForgetExportRequest::TPtr& ev, const TAc
 
 void TSchemeShard::Handle(TEvExport::TEvListExportsRequest::TPtr& ev, const TActorContext& ctx) {
     Execute(CreateTxListExports(ev), ctx);
-}
-
-void TSchemeShard::Handle(TEvPrivate::TEvExportSchemeUploadResult::TPtr& ev, const TActorContext& ctx) {
-    Execute(CreateTxProgressExport(ev), ctx);
 }
 
 void TSchemeShard::ResumeExports(const TVector<ui64>& exportIds, const TActorContext& ctx) {
