+++ ydb/core/tx/schemeshard/schemeshard_export__create.cpp
@@ -1,9 +1,8 @@
 #include "schemeshard_xxport__tx_base.h"
 #include "schemeshard_xxport__helpers.h"
-#include "schemeshard_export.h"
 #include "schemeshard_export_flow_proposals.h"
 #include "schemeshard_export_helpers.h"
-#include "schemeshard_export_scheme_uploader.h"
+#include "schemeshard_export.h"
 #include "schemeshard_audit_log.h"
 #include "schemeshard_impl.h"
 
@@ -204,7 +203,7 @@ private:
                     .IsResolved()
                     .NotDeleted()
                     .NotUnderDeleting()
-                    .IsSupportedInExports()
+                    .IsTable()
                     .FailOnRestrictedCreateInTempZone();
 
                 if (!checks) {
@@ -213,7 +212,7 @@ private:
                 }
             }
 
-            exportInfo->Items.emplace_back(item.source_path(), path.Base()->PathId, path->PathType);
+            exportInfo->Items.emplace_back(item.source_path(), path.Base()->PathId);
             exportInfo->PendingItems.push_back(itemIdx);
         }
 
@@ -231,7 +230,6 @@ struct TSchemeShard::TExport::TTxProgress: public TSchemeShard::TXxport::TTxBase
     ui64 Id;
     TEvTxAllocatorClient::TEvAllocateResult::TPtr AllocateResult = nullptr;
     TEvSchemeShard::TEvModifySchemeTransactionResult::TPtr ModifyResult = nullptr;
-    TEvPrivate::TEvExportSchemeUploadResult::TPtr SchemeUploadResult = nullptr;
     TTxId CompletedTxId = InvalidTxId;
 
     explicit TTxProgress(TSelf* self, ui64 id)
@@ -252,12 +250,6 @@ struct TSchemeShard::TExport::TTxProgress: public TSchemeShard::TXxport::TTxBase
     {
     }
 
-    explicit TTxProgress(TSelf* self, TEvPrivate::TEvExportSchemeUploadResult::TPtr& ev)
-        : TXxport::TTxBase(self)
-        , SchemeUploadResult(ev)
-    {
-    }
-
     explicit TTxProgress(TSelf* self, TTxId completedTxId)
         : TXxport::TTxBase(self)
         , CompletedTxId(completedTxId)
@@ -275,8 +267,6 @@ struct TSchemeShard::TExport::TTxProgress: public TSchemeShard::TXxport::TTxBase
             OnAllocateResult(txc, ctx);
         } else if (ModifyResult) {
             OnModifyResult(txc, ctx);
-        } else if (SchemeUploadResult) {
-            OnSchemeUploadResult(txc, ctx);
         } else if (CompletedTxId) {
             OnNotifyResult(txc, ctx);
         } else {
@@ -300,22 +290,16 @@ private:
         Send(Self->SelfId(), MkDirPropose(Self, txId, exportInfo));
     }
 
-    bool CopyTables(TExportInfo::TPtr exportInfo, TTxId txId) {
+    void CopyTables(TExportInfo::TPtr exportInfo, TTxId txId) {
         LOG_I("TExport::TTxProgress: CopyTables propose"
             << ": info# " << exportInfo->ToString()
             << ", txId# " << txId);
 
         Y_ABORT_UNLESS(exportInfo->WaitTxId == InvalidTxId);
-        if (AnyOf(exportInfo->Items, [](const TExportInfo::TItem& item) {
-            return item.SourcePathType == NKikimrSchemeOp::EPathTypeTable;
-        })) {
-            Send(Self->SelfId(), CopyTablesPropose(Self, txId, exportInfo));
-            return true;
-        }
-        return false;
+        Send(Self->SelfId(), CopyTablesPropose(Self, txId, exportInfo));
     }
 
-    void TransferData(TExportInfo::TPtr exportInfo, ui32 itemIdx, TTxId txId, const TActorContext& ctx) {
+    void TransferData(TExportInfo::TPtr exportInfo, ui32 itemIdx, TTxId txId) {
         Y_ABORT_UNLESS(itemIdx < exportInfo->Items.size());
         auto& item = exportInfo->Items.at(itemIdx);
 
@@ -327,17 +311,7 @@ private:
             << ", txId# " << txId);
 
         Y_ABORT_UNLESS(item.WaitTxId == InvalidTxId);
-        if (item.SourcePathType == NKikimrSchemeOp::EPathTypeTable) {
-            Send(Self->SelfId(), BackupPropose(Self, txId, exportInfo, itemIdx));
-        } else if (item.SourcePathType == NKikimrSchemeOp::EPathTypeView) {
-            Ydb::Export::ExportToS3Settings exportSettings;
-            Y_ABORT_UNLESS(exportSettings.ParseFromString(exportInfo->Settings));
-            const auto databaseRoot = TStringBuilder() << '/' << JoinSeq('/', Self->RootPathElements);
-
-            item.SchemeUploader = ctx.Register(CreateSchemeUploader(
-                Self->SelfId(), exportInfo->Id, itemIdx, item.SourcePathId, exportSettings, databaseRoot
-            ));
-        }
+        Send(Self->SelfId(), BackupPropose(Self, txId, exportInfo, itemIdx));
     }
 
     bool CancelTransferring(TExportInfo::TPtr exportInfo, ui32 itemIdx) {
@@ -646,7 +620,7 @@ private:
         }
     }
 
-    void OnAllocateResult(TTransactionContext& txc, const TActorContext& ctx) {
+    void OnAllocateResult(TTransactionContext&, const TActorContext&) {
         Y_ABORT_UNLESS(AllocateResult);
 
         const auto txId = TTxId(AllocateResult->Get()->TxIds.front());
@@ -677,27 +651,13 @@ private:
             break;
 
         case EState::CopyTables:
-            if (!CopyTables(exportInfo, txId)) {
-                // none of the items is a table
-                NIceDb::TNiceDb db(txc.DB);
-
-                for (ui32 itemIdx : xrange(exportInfo->Items.size())) {
-                    exportInfo->Items[itemIdx].State = EState::Transferring;
-                    Self->PersistExportItemState(db, exportInfo, itemIdx);
-
-                    AllocateTxId(exportInfo, itemIdx);
-                }
-
-                exportInfo->State = EState::Transferring;
-                Self->PersistExportState(db, exportInfo);
-                return;
-            }
+            CopyTables(exportInfo, txId);
             break;
 
         case EState::Transferring:
             if (exportInfo->PendingItems) {
                 itemIdx = popPendingItemIdx(exportInfo->PendingItems);
-                TransferData(exportInfo, itemIdx, txId, ctx);
+                TransferData(exportInfo, itemIdx, txId);
             } else {
                 return;
             }
@@ -910,71 +870,6 @@ private:
         SubscribeTx(txId);
     }
 
-    void OnSchemeUploadResult(TTransactionContext& txc, const TActorContext& ctx) {
-        Y_ABORT_UNLESS(SchemeUploadResult);
-
-        LOG_D("TExport::TTxProgress: OnSchemeUploadResult"
-            << ": success# " << SchemeUploadResult->Get()->Success
-            << ", error# " << SchemeUploadResult->Get()->Error
-        );
-
-        const auto exportId = SchemeUploadResult->Get()->ExportId;
-        auto exportInfo = Self->Exports.Value(exportId, nullptr);
-        if (!exportInfo) {
-            LOG_E("TExport::TTxProgress: OnSchemeUploadResult received unknown export id"
-                << ": id# " << exportId
-            );
-            return;
-        }
-
-        ui32 itemIdx = SchemeUploadResult->Get()->ItemIdx;
-        if (itemIdx >= exportInfo->Items.size()) {
-            LOG_E("TExport::TTxProgress: OnSchemeUploadResult item index out of range"
-                << ": id# " << exportId
-                << ", item index# " << itemIdx
-                << ", number of items# " << exportInfo->Items.size()
-            );
-            return;
-        }
-
-        NIceDb::TNiceDb db(txc.DB);
-
-        auto& item = exportInfo->Items[itemIdx];
-        item.SchemeUploader = TActorId();
-
-        if (!SchemeUploadResult->Get()->Success) {
-            item.State = EState::Cancelled;
-            item.Issue = SchemeUploadResult->Get()->Error;
-            Self->PersistExportItemState(db, exportInfo, itemIdx);
-
-            if (!exportInfo->IsInProgress()) {
-                return;
-            }
-
-            Cancel(exportInfo, itemIdx, "unsuccessful scheme upload");
-
-            Self->PersistExportState(db, exportInfo);
-            return SendNotificationsIfFinished(exportInfo);
-        }
-
-        if (exportInfo->State == EState::Transferring) {
-            item.State = EState::Done;
-            Self->PersistExportItemState(db, exportInfo, itemIdx);
-
-            if (AllOf(exportInfo->Items, &TExportInfo::TItem::IsDone)) {
-                exportInfo->State = EState::Done;
-                exportInfo->EndTime = TAppData::TimeProvider->Now();
-
-                Self->PersistExportState(db, exportInfo);
-                SendNotificationsIfFinished(exportInfo);
-
-                if (exportInfo->IsFinished()) {
-                    AuditLogExportEnd(*exportInfo.Get(), Self);
-                }
-            }
-        }
-    }
-
     void OnNotifyResult(TTransactionContext& txc, const TActorContext&) {
         Y_ABORT_UNLESS(CompletedTxId);
         LOG_D("TExport::TTxProgress: OnNotifyResult"
@@ -1053,17 +948,14 @@ private:
             item.State = EState::Done;
             item.WaitTxId = InvalidTxId;
 
-            bool itemHasIssues = false;
-            if (item.SourcePathType == NKikimrSchemeOp::EPathTypeTable) {
-                if (const auto issue = GetIssues(ItemPathId(Self, exportInfo, itemIdx), txId)) {
-                    item.Issue = *issue;
-                    Cancel(exportInfo, itemIdx, "issues during backing up");
-                    itemHasIssues = true;
+            if (const auto issue = GetIssues(ItemPathId(Self, exportInfo, itemIdx), txId)) {
+                item.Issue = *issue;
+                Cancel(exportInfo, itemIdx, "issues during backing up");
+            } else {
+                if (AllOf(exportInfo->Items, &TExportInfo::TItem::IsDone)) {
+                    exportInfo->State = EState::Done;
+                    exportInfo->EndTime = TAppData::TimeProvider->Now();
                 }
-            }
-            if (!itemHasIssues && AllOf(exportInfo->Items, &TExportInfo::TItem::IsDone)) {
-                exportInfo->State = EState::Done;
-                exportInfo->EndTime = TAppData::TimeProvider->Now();
             }
 
             Self->PersistExportItemState(db, exportInfo, itemIdx);
@@ -1121,10 +1013,6 @@ ITransaction* TSchemeShard::CreateTxProgressExport(TEvTxAllocatorClient::TEvAllo
 }
 
 ITransaction* TSchemeShard::CreateTxProgressExport(TEvSchemeShard::TEvModifySchemeTransactionResult::TPtr& ev) {
-    return new TExport::TTxProgress(this, ev);
-}
-
-ITransaction* TSchemeShard::CreateTxProgressExport(TEvPrivate::TEvExportSchemeUploadResult::TPtr& ev) {
     return new TExport::TTxProgress(this, ev);
 }
