+++ ydb/core/tx/schemeshard/schemeshard_export_flow_proposals.cpp
@@ -59,9 +59,6 @@ THolder<TEvSchemeShard::TEvModifySchemeTransaction> CopyTablesPropose(
 
     for (ui32 itemIdx : xrange(exportInfo->Items.size())) {
         const auto& item = exportInfo->Items.at(itemIdx);
-        if (item.SourcePathType != NKikimrSchemeOp::EPathTypeTable) {
-            continue;
-        }
 
         auto& desc = *copyTables.Add();
         desc.SetSrcPath(item.SourcePathName);
