+++ ydb/core/tx/schemeshard/schemeshard_info_types.h
@@ -2664,20 +2664,17 @@ struct TExportInfo: public TSimpleRefCount<TExportInfo> {
 
         TString SourcePathName;
         TPathId SourcePathId;
-        NKikimrSchemeOp::EPathType SourcePathType;
 
         EState State = EState::Waiting;
         ESubState SubState = ESubState::AllocateTxId;
         TTxId WaitTxId = InvalidTxId;
-        TActorId SchemeUploader;
         TString Issue;
 
         TItem() = default;
 
-        explicit TItem(const TString& sourcePathName, const TPathId sourcePathId, NKikimrSchemeOp::EPathType sourcePathType)
+        explicit TItem(const TString& sourcePathName, const TPathId sourcePathId)
             : SourcePathName(sourcePathName)
             , SourcePathId(sourcePathId)
-            , SourcePathType(sourcePathType)
         {
         }
