+++ ydb/core/tx/schemeshard/schemeshard_path.h
@@ -104,7 +104,6 @@ public:
         const TChecker& FailOnRestrictedCreateInTempZone(bool allowCreateInTemporaryDir = false, EStatus status = EStatus::StatusPreconditionFailed) const;
         const TChecker& IsResourcePool(EStatus status = EStatus::StatusNameConflict) const;
         const TChecker& IsBackupCollection(EStatus status = EStatus::StatusNameConflict) const;
-        const TChecker& IsSupportedInExports(EStatus status = EStatus::StatusNameConflict) const;
     };
 
 public:
