+++ ydb/core/tx/schemeshard/schemeshard_schema.h
@@ -1209,7 +1209,6 @@ struct Schema : NIceDb::Schema {
         struct BackupTxId : Column<6, NScheme::NTypeIds::Uint64> { using Type = TTxId; };
         struct Issue : Column<7, NScheme::NTypeIds::Utf8> {};
         struct SourceOwnerPathId : Column<8, NScheme::NTypeIds::Uint64> { using Type = TOwnerId; };
-        struct SourcePathType : Column<9, NScheme::NTypeIds::Uint32> { using Type = NKikimrSchemeOp::EPathType; static constexpr Type Default = NKikimrSchemeOp::EPathTypeTable; };
 
         using TKey = TableKey<ExportId, Index>;
         using TColumns = TableColumns<
@@ -1220,8 +1219,7 @@ struct Schema : NIceDb::Schema {
             State,
             BackupTxId,
             Issue,
-            SourceOwnerPathId,
-            SourcePathType
+            SourceOwnerPathId
         >;
     };
