+++ ydb/core/tx/schemeshard/ya.make
@@ -306,7 +306,6 @@ PEERDIR(
     ydb/library/login
     ydb/library/login/protos
     ydb/library/protobuf_printer
-    ydb/public/lib/ydb_cli/dump/util
     yql/essentials/minikql
     yql/essentials/providers/common/proto
     ydb/services/bg_tasks
@@ -318,12 +317,10 @@ YQL_LAST_ABI_VERSION()
 
 IF (OS_WINDOWS)
     SRCS(
-        schemeshard_export_scheme_uploader_fallback.cpp
         schemeshard_import_scheme_getter_fallback.cpp
     )
 ELSE()
     SRCS(
-        schemeshard_export_scheme_uploader.cpp
         schemeshard_import_scheme_getter.cpp
     )
 ENDIF()
