+++ ydb/public/lib/ydb_cli/dump/util/view_utils.h
@@ -2,25 +2,10 @@
 
 namespace NYdb::NDump {
 
+bool ValidateViewQuery(const TString& query, NYql::TIssues& issues);
+
 TString RewriteAbsolutePath(TStringBuf path, TStringBuf backupRoot, TStringBuf restoreRoot);
 
 bool RewriteTableRefs(TString& scheme, TStringBuf backupRoot, TStringBuf restoreRoot, NYql::TIssues& issues);
-
-struct TViewQuerySplit {
-    TString ContextRecreation;
-    TString Select;
-};
-
-TViewQuerySplit SplitViewQuery(TStringInput query);
-
-// returns void, because the validation is non-blocking
-void ValidateViewQuery(const TString& query, const TString& dbPath, NYql::TIssues& issues);
-
-bool Format(const TString& query, TString& formattedQuery, NYql::TIssues& issues);
-
-TString BuildCreateViewQuery(
-    const TString& name, const TString& dbPath, const TString& viewQuery, const TString& backupRoot,
-    NYql::TIssues& issues
-);
 
 }
