+++ ydb/tests/functional/scheme_tests/canondata/tablet_scheme_tests.TestTabletSchemes.test_tablet_schemes_flat_schemeshard_/flat_schemeshard.schema
@@ -3205,11 +3205,6 @@
                 "ColumnId": 8,
                 "ColumnName": "SourceOwnerPathId",
                 "ColumnType": "Uint64"
-            },
-            {
-                "ColumnId": 9,
-                "ColumnName": "SourcePathType",
-                "ColumnType": "Uint32"
             }
         ],
         "ColumnsDropped": [],
@@ -3223,8 +3218,7 @@
                     5,
                     6,
                     7,
-                    8,
-                    9
+                    8
                 ],
                 "RoomID": 0,
                 "Codec": 0,
